+++ src/AppBundle/Controller/AuthenticationController.php
@@ -148,7 +148,9 @@ class AuthenticationController extends Controller
 
         $logger->info('Return authentication page with QR code');
 
-        return $this->render('AppBundle:default:authentication.html.twig', []);
+        return $this->render('AppBundle:default:authentication.html.twig', [
+            'authenticateUrl' => $this->tiqrService->authenticationUrl()
+        ]);
     }
 
     /**
@@ -164,10 +166,6 @@ class AuthenticationController extends Controller
             return $this->refreshAuthenticationPage();
         }
 
-        if ($this->authenticationChallengeIsExpired()) {
-            return $this->timeoutNeedsManualRetry();
-        }
-
         $isAuthenticated = $this->tiqrService->isAuthenticated();
 
         if ($isAuthenticated) {
@@ -175,6 +173,11 @@ class AuthenticationController extends Controller
 
             return $this->refreshAuthenticationPage();
         }
+
+        if ($this->authenticationChallengeIsExpired()) {
+            return $this->timeoutNeedsManualRetry();
+        }
+
         $this->logger->info('Send json response is not authenticated');
 
         return $this->scheduleNextPollOnAuthenticationPage();