+++ src/AppBundle/Resources/javascript/Client/NotificationClient.ts
@@ -7,6 +7,12 @@ export class NotificationClient {
    * Request status form the API.
    */
   public send(callback: (result: string) => void, errorCallback: (error: unknown) => void) {
+    jQuery.ajaxSetup(
+      {
+        // Wait max 10 seconds.
+        timeout: 1000 * 10,
+      },
+    );
     jQuery.post(this.apiUrl, callback).fail(errorCallback);
   }
 }